requirements.txt
Flask==3.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
flask-cors==4.0.0
Pillow==10.0.1
pytesseract==0.3.10
opencv-python==4.8.1.78
numpy==1.24.3

